+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/Account.kt
@@ -24,7 +24,6 @@ import androidx.compose.runtime.Stable
 import com.vitorpamplona.amethyst.Amethyst
 import com.vitorpamplona.amethyst.BuildConfig
 import com.vitorpamplona.amethyst.LocalPreferences
-import com.vitorpamplona.amethyst.R
 import com.vitorpamplona.amethyst.commons.audio.VisualizerStyle
 import com.vitorpamplona.amethyst.commons.connectedApps.nip46.InMemoryNip46ClientStore
 import com.vitorpamplona.amethyst.commons.connectedApps.nip46.Nip46ClientStore
@@ -60,11 +59,6 @@ import com.vitorpamplona.amethyst.commons.model.nip85TrustedAssertions.ContactCa
 import com.vitorpamplona.amethyst.commons.model.nip85TrustedAssertions.ContactCardsState
 import com.vitorpamplona.amethyst.commons.model.nip85TrustedAssertions.TrustProviderListDecryptionCache
 import com.vitorpamplona.amethyst.commons.model.privateChats.hasEncryptedContent
-import com.vitorpamplona.amethyst.commons.onchain.OnchainZapSendError
-import com.vitorpamplona.amethyst.commons.onchain.OnchainZapSendResult
-import com.vitorpamplona.amethyst.commons.onchain.OnchainZapSendStage
-import com.vitorpamplona.amethyst.commons.onchain.OnchainZapSender
-import com.vitorpamplona.amethyst.commons.onchain.OnchainZapShare
 import com.vitorpamplona.amethyst.commons.relayauth.RelayAuthCustomToggles
 import com.vitorpamplona.amethyst.commons.relayauth.RelayAuthPermissionStore
 import com.vitorpamplona.amethyst.commons.richtext.RichTextParser
@@ -254,12 +248,6 @@ import com.vitorpamplona.quartz.nip37Drafts.DraftWrapEvent
 import com.vitorpamplona.quartz.nip42RelayAuth.RelayAuthEvent
 import com.vitorpamplona.quartz.nip47WalletConnect.Nip47WalletConnect
 import com.vitorpamplona.quartz.nip47WalletConnect.events.NwcInfoEvent
-import com.vitorpamplona.quartz.nip47WalletConnect.rpc.IErrorResponseLike
-import com.vitorpamplona.quartz.nip47WalletConnect.rpc.NwcMethod
-import com.vitorpamplona.quartz.nip47WalletConnect.rpc.PayMethod
-import com.vitorpamplona.quartz.nip47WalletConnect.rpc.PaySuccessResponse
-import com.vitorpamplona.quartz.nip47WalletConnect.rpc.Request
-import com.vitorpamplona.quartz.nip47WalletConnect.rpc.Response
 import com.vitorpamplona.quartz.nip51Lists.bookmarkList.BookmarkListEvent
 import com.vitorpamplona.quartz.nip51Lists.bookmarkList.tags.AddressBookmark
 import com.vitorpamplona.quartz.nip56Reports.ReportEvent
@@ -319,8 +307,6 @@ import com.vitorpamplona.quartz.nipA0VoiceMessages.BaseVoiceEvent
 import com.vitorpamplona.quartz.nipA0VoiceMessages.VoiceEvent
 import com.vitorpamplona.quartz.nipA0VoiceMessages.VoiceReplyEvent
 import com.vitorpamplona.quartz.nipB0WebBookmarks.WebBookmarkEvent
-import com.vitorpamplona.quartz.nipB1Bolt12Zaps.builder.Bolt12ZapBuilder
-import com.vitorpamplona.quartz.nipB1Bolt12Zaps.verify.Bolt12ZapValidation
 import com.vitorpamplona.quartz.nipC7Chats.ChatEvent
 import com.vitorpamplona.quartz.utils.DualCase
 import com.vitorpamplona.quartz.utils.Log
@@ -346,8 +332,6 @@ import java.math.BigDecimal
 import kotlin.coroutines.cancellation.CancellationException
 import com.vitorpamplona.quartz.experimental.nip95.header.thumbhash as nip95thumbhash
 import com.vitorpamplona.quartz.experimental.profileGallery.thumbhash as galleryThumbhash
-
-private const val ONCHAIN_BACKEND_NOT_CONFIGURED = "Bitcoin chain backend is not configured"
 
 @OptIn(DelicateCoroutinesApi::class)
 @Stable
@@ -725,6 +709,9 @@ class Account(
 
     /** NIP-29 relay-group + Buzz workspace orchestration. */
     val relayGroups = AccountRelayGroupActions(this)
+
+    /** Zap/payment orchestration (NIP-57, NWC, BOLT12, onchain). */
+    val zaps = AccountZapActions(this)
 
     /**
      * Relay routing + sign-and-publish choke point: computes which relays an event
@@ -1337,278 +1324,6 @@ class Account(
      */
     fun consumeLabelEvent(event: Event) {
         cache.justConsumeMyOwnEvent(event)
-    }
-
-    suspend fun createZapRequestFor(
-        event: Event,
-        pollOption: Int?,
-        message: String = "",
-        zapType: LnZapEvent.ZapType,
-        toUser: User?,
-        additionalRelays: Set<NormalizedRelayUrl>? = null,
-        amountMillisats: Long? = null,
-        lnurl: String? = null,
-    ) = LnZapRequestEvent.create(
-        zappedEvent = event,
-        relays = nip65RelayList.inboxFlow.value + (additionalRelays ?: emptySet()),
-        signer = signer,
-        pollOption = pollOption,
-        message = message,
-        zapType = zapType,
-        toUserPubHex = toUser?.pubkeyHex,
-        amountMillisats = amountMillisats,
-        lnurl = lnurl,
-    )
-
-    suspend fun calculateIfNoteWasZappedByAccount(
-        zappedNote: Note?,
-        afterTimeInSeconds: Long,
-    ): Boolean = zappedNote?.isZappedBy(userProfile(), afterTimeInSeconds, this) == true
-
-    suspend fun calculateZappedAmount(zappedNote: Note): BigDecimal = zappedNote.zappedAmountWithNWCPayments(nip47SignerState)
-
-    suspend fun sendNwcRequest(
-        request: Request,
-        onResponse: (Response?) -> Unit,
-    ) {
-        val (event, relay) = nip47SignerState.sendNwcRequest(request, onResponse)
-        client.publish(event, setOf(relay))
-    }
-
-    suspend fun sendNwcRequestToWallet(
-        walletUri: Nip47WalletConnect.Nip47URINorm,
-        request: Request,
-        onResponse: (Response?) -> Unit,
-    ): HexKey {
-        val (event, relay) = nip47SignerState.sendNwcRequestToWallet(walletUri, request, onResponse)
-        client.publish(event, setOf(relay))
-        return event.id
-    }
-
-    /**
-     * Number of spoofed (wrong-author) NIP-47 replies that have arrived for
-     * the given request id. 0 if the request is unknown or already resolved.
-     */
-    fun nwcSpoofAttempts(requestId: HexKey): Int = LocalCache.paymentTracker.spoofAttemptsFor(requestId)
-
-    /**
-     * Removes a pending NIP-47 request from the tracker. Call this when the
-     * UI gives up waiting (timeout) so the entry doesn't stick around.
-     */
-    fun cleanupNwcRequest(requestId: HexKey) = LocalCache.paymentTracker.cleanup(requestId)
-
-    suspend fun sendZapPaymentRequestFor(
-        bolt11: String,
-        zappedNote: Note?,
-        onResponse: (Response?) -> Unit,
-    ) {
-        val (event, relay) = nip47SignerState.sendZapPaymentRequestFor(bolt11, zappedNote, onResponse)
-        client.publish(event, setOf(relay))
-    }
-
-    /**
-     * True when the default NWC wallet advertises the nwc#2 `pay` method — the rail a
-     * BOLT12 zap needs to obtain a payer proof. Read from the wallet's cached kind:13194
-     * info event (its capability advertisement), which [NwcSignerState] already refreshes
-     * on wallet change. A missing/unfetched info event reads as false, so the zap path
-     * falls back to lightning rather than attempting a `pay` the wallet can't honor.
-     */
-    fun defaultWalletSupportsBolt12Pay(): Boolean {
-        val uri = nip47SignerState.defaultWalletUri.value ?: return false
-        return nip47SignerState.infoCache?.current(uri)?.supportsMethod(NwcMethod.PAY) == true
-    }
-
-    /**
-     * Sends a NIP-B1 BOLT12 zap to [recipientPubKey] over the default NWC wallet.
-     *
-     * Signs a kind 9737 intent, pays [offer] via the nwc#2 `pay` method with the
-     * intent-bound `payer_note`, then — only if the wallet returns a payer proof that
-     * validates — builds, self-consumes, and publishes the kind 9736 zap. Validation
-     * is the fail-safe: a wallet that drops or misroutes the note yields a proof that
-     * fails the binding check, so no invalid receipt is ever published (the payment
-     * still happened; [onError] reports "paid, no receipt"). [zappedEvent] is null for
-     * a profile zap. Requires an NWC wallet (see [hasNwcWallet]); BOLT12 zaps have no
-     * external-wallet or LNURL fallback because only NWC returns the proof.
-     */
-    suspend fun sendBolt12Zap(
-        zappedEvent: Event?,
-        recipientPubKey: HexKey,
-        offer: String,
-        amountMillisats: Long,
-        message: String,
-        zapType: LnZapEvent.ZapType,
-        // (messageResId, detail) — the caller localizes; detail carries a wallet error, if any.
-        onError: (Int, String?) -> Unit,
-        onProcessed: () -> Unit,
-    ) {
-        // NONZAP means "pay, but publish no receipt" — settle the offer without binding
-        // a zap intent or emitting a 9736, matching the privacy of a bolt11 NONZAP.
-        if (zapType == LnZapEvent.ZapType.NONZAP) {
-            sendNwcRequest(PayMethod.create("bitcoin:?lno=$offer", amountMillisats)) { response ->
-                scope.launch {
-                    if (response is IErrorResponseLike) onError(R.string.bolt12_payment_failed, response.errorMessage())
-                    onProcessed()
-                }
-            }
-            return
-        }
-
-        val anonymous = zapType == LnZapEvent.ZapType.ANONYMOUS
-        // The 9737 intent and the 9736 zap MUST be signed by the same key. An anonymous
-        // zap uses a fresh ephemeral key so it carries no `P` tag and isn't traceable.
-        val zapSigner = if (anonymous) NostrSignerInternal(KeyPair()) else signer
-
-        val intent =
-            if (zappedEvent == null) {
-                Bolt12ZapBuilder.buildProfileIntent(zapSigner, recipientPubKey, amountMillisats, offer, message)
-            } else {
-                Bolt12ZapBuilder.buildIntent(zapSigner, recipientPubKey, amountMillisats, offer, EventHintBundle(zappedEvent), message)
-            }
-
-        val payerNote = Bolt12ZapBuilder.payerNote(intent)
-
-        sendNwcRequest(PayMethod.create("bitcoin:?lno=$offer", amountMillisats, payerNote)) { response ->
-            scope.launch {
-                // try/finally so a failure while assembling/publishing the receipt (e.g. a
-                // remote signer error) still steps progress and surfaces an error, instead
-                // of vanishing as an uncaught coroutine exception. The payment already
-                // settled at this point, so such a failure means "paid, no receipt".
-                try {
-                    when (response) {
-                        is PaySuccessResponse -> {
-                            val proof = response.result?.payer_proof
-                            if (proof.isNullOrBlank()) {
-                                onError(R.string.bolt12_zap_paid_no_receipt, null)
-                            } else {
-                                val zap = Bolt12ZapBuilder.buildZap(zapSigner, intent, proof, anonymous)
-                                if (cache.bolt12ZapValidator.validate(zap, verifyEventSignature = false) is Bolt12ZapValidation.Valid) {
-                                    cache.justConsumeMyOwnEvent(zap)
-                                    client.publish(zap, computeRelayListToBroadcast(zap))
-                                } else {
-                                    onError(R.string.bolt12_zap_invalid_receipt, null)
-                                }
-                            }
-                        }
-
-                        is IErrorResponseLike -> onError(R.string.bolt12_payment_failed, response.errorMessage())
-
-                        else -> onError(R.string.bolt12_zap_paid_no_receipt, null)
-                    }
-                } catch (e: CancellationException) {
-                    throw e
-                } catch (e: Exception) {
-                    Log.w("Account", "BOLT12 zap receipt assembly failed after payment", e)
-                    onError(R.string.bolt12_zap_paid_no_receipt, null)
-                } finally {
-                    onProcessed()
-                }
-            }
-        }
-    }
-
-    suspend fun createZapRequestFor(
-        user: User,
-        message: String = "",
-        zapType: LnZapEvent.ZapType,
-        amountMillisats: Long? = null,
-        lnurl: String? = null,
-    ): LnZapRequestEvent {
-        val zapRequest =
-            LnZapRequestEvent.create(
-                userHex = user.pubkeyHex,
-                relays = nip65RelayList.inboxFlow.value + (user.inboxRelays() ?: emptyList()),
-                signer = signer,
-                message = message,
-                zapType = zapType,
-                amountMillisats = amountMillisats,
-                lnurl = lnurl,
-            )
-
-        cache.justConsumeMyOwnEvent(zapRequest)
-        return zapRequest
-    }
-
-    private fun onchainBackendNotConfigured() =
-        OnchainZapSendResult.Failure(
-            OnchainZapSendStage.LOADING_UTXOS,
-            OnchainZapSendError.BACKEND_NOT_CONFIGURED,
-            ONCHAIN_BACKEND_NOT_CONFIGURED,
-        )
-
-    /**
-     * Send a NIP-BC onchain zap: build a Bitcoin transaction paying the recipient's
-     * derived Taproot address, sign it, broadcast it, and publish the kind:8333
-     * zap receipt. Pass [zappedEvent] to attribute the zap to a specific event, or
-     * leave it null for a profile zap.
-     */
-    suspend fun sendOnchainZap(
-        recipientPubKey: HexKey,
-        amountSats: Long,
-        feeRateSatPerVByte: Double,
-        comment: String = "",
-        zappedEvent: EventHintBundle<out Event>? = null,
-    ): OnchainZapSendResult {
-        val backend =
-            cache.onchainBackend
-                ?: return onchainBackendNotConfigured()
-        return OnchainZapSender.send(
-            backend = backend,
-            signer = signer,
-            senderPubKey = signer.pubKey,
-            recipientPubKey = recipientPubKey,
-            amountSats = amountSats,
-            feeRateSatPerVByte = feeRateSatPerVByte,
-            comment = comment,
-            zappedEvent = zappedEvent,
-        ) { template -> signAndComputeBroadcast(template) }
-    }
-
-    /**
-     * Pay an explicit Bitcoin address (e.g. a profile's NIP-A3 `bitcoin`
-     * payment target) from the NIP-BC Taproot wallet. A plain wallet send —
-     * no kind:8333 receipt is published. See [OnchainZapSender.sendToAddress].
-     */
-    suspend fun sendOnchainToAddress(
-        recipientAddress: String,
-        amountSats: Long,
-        feeRateSatPerVByte: Double,
-    ): OnchainZapSendResult {
-        val backend =
-            cache.onchainBackend
-                ?: return onchainBackendNotConfigured()
-        return OnchainZapSender.sendToAddress(
-            backend = backend,
-            signer = signer,
-            senderPubKey = signer.pubKey,
-            recipientAddress = recipientAddress,
-            amountSats = amountSats,
-            feeRateSatPerVByte = feeRateSatPerVByte,
-        )
-    }
-
-    /**
-     * Send a NIP-BC onchain split zap: a single Bitcoin transaction paying
-     * each recipient their precomputed share, plus one kind:8333 receipt per
-     * recipient. See [OnchainZapSender.sendSplit] for failure semantics.
-     */
-    suspend fun sendOnchainZapWithSplits(
-        recipients: List<OnchainZapShare>,
-        feeRateSatPerVByte: Double,
-        comment: String = "",
-        zappedEvent: EventHintBundle<out Event>? = null,
-    ): OnchainZapSendResult {
-        val backend =
-            cache.onchainBackend
-                ?: return onchainBackendNotConfigured()
-        return OnchainZapSender.sendSplit(
-            backend = backend,
-            signer = signer,
-            senderPubKey = signer.pubKey,
-            recipients = recipients,
-            feeRateSatPerVByte = feeRateSatPerVByte,
-            comment = comment,
-            zappedEvent = zappedEvent,
-        ) { template -> signAndComputeBroadcast(template) }
     }
 
     suspend fun report(
